Enhancement: Add newPatient type (#12159)
