+++ src/pages/Organization/OrganizationPatients.tsx
@@ -18,7 +18,7 @@ import useFilters from "@/hooks/useFilters";
 
 import query from "@/Utils/request/query";
 import { formatDateTime, relativeTime } from "@/Utils/utils";
-import { Patient } from "@/types/emr/newPatient";
+import { Patient } from "@/types/emr/patient";
 import { Organization } from "@/types/organization/organization";
 import organizationApi from "@/types/organization/organizationApi";
