+++ src/types/resourceRequest/resourceRequest.ts
@@ -1,7 +1,7 @@
 import { FacilityModel } from "@/components/Facility/models";
 import { UserBareMinimum } from "@/components/Users/models";
 
-import { PatientModel } from "@/types/emr/patient";
+import { Patient } from "@/types/emr/patient";
 import { UserBase } from "@/types/user/user";
 
 export interface ResourceRequest {
@@ -23,7 +23,7 @@ export interface ResourceRequest {
   updated_by: UserBase;
   created_date: string;
   modified_date: string;
-  related_patient: PatientModel | null;
+  related_patient: Patient | null;
 }
 
 export interface CreateResourceRequest {
